Move response logics to the ResponseHandler module
