+++ lib/http/wrapper.rb
@@ -8,6 +8,7 @@ require "http/wrapper/api_exceptions"
 require "http/wrapper/configuration"
 require "http/wrapper/error_mapping"
 require "http/wrapper/request_handler"
+require "http/wrapper/response_handler"
 require "http/wrapper/client"
 
 module Http
